Basic caching support
Just caches in memory right now (will likely replace that whole
mechanism to always cache to the HDD... not sure how important a memory
cache is going to be overall, but we'll see).

Seems to work like a charm though! Really low CPU use both when doing
upstream requests and when just full-filling from the cache. Not
entirely sure multithreading is going to even be needed for our usage.
In theory the locking is there though.