Reorg to avoid dynamic memory allocation
This seems to pretty much entirely bypass the GC issues. Slightly ugly,
but goes plleeenntty fast on an SSD at least (suspecting you just hit
the OS file cache as well, if not the disk cache) so this is the first
workable commit for now.

Will think of ways to clean things up better and avoid duplication.