Fix a few robustness cases for the BYPASS path
Broadcasting and chat and similar now go through the same DNS hosts as
steam content servers, so we need to fix some robustness issues with the
bypass path.

Probably worth specializing it a bit further even, at least in terms of
the response writing (writeRawBody), but this at least restores
functionality for now.